requirements.txt
speedtest-cli==2.1.3
rich==13.7.0
click==8.1.7

